% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2.R
\name{d2}
\alias{d2}
\title{d2 Function for SPC Charts}
\usage{
d2(n)
}
\arguments{
\item{n}{A numeric vector of length at least 1 and all values integers of 1 or larger}
}
\value{
A numeric vector
}
\description{
The d2 function is used as a weighting variable used in R charts, generally
when the number of measurements in a run is 10 or less.
}
\examples{
n <- c(2, 4, 3, 2)
d2(n)
}
