% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lithograph.R
\docType{data}
\name{lithograph}
\alias{lithograph}
\title{The width of a line on a silicon wafer.}
\format{
A data frame with 450 rows and 5 variables:
\describe{
\item{CASSETTE}{Cassette number}
\item{WAFER}{Wafer number in a given cassette}
\item{SITE}{ Site on the wafer: 1=Top, 2=Left, 3=Center, 4=Right, 5=Bottom}
\item{LINEWIDT}{Width of the line on the wafer}
\item{RUNSEQ}{Sequence number (1-450)}
}
}
\source{
\url{https://www.itl.nist.gov/div898/handbook/datasets/LITHOGRA.DAT}
}
\usage{
lithograph
}
\description{
This data is taken from the
\href{https://www.itl.nist.gov/div898/handbook/}{NIST/Sematech Engineering Statistics Handbook}.
The width of a line on a silicon wafer is measured for
several factor variables.
}
\keyword{datasets}
